pyproject.toml
[tool.poetry]
name = "azure-rm-proxy-server"
version = "0.1.19"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    {include = "azure_rm_proxy"},
    {include = "azure_rm_client"},
    {include = "azure_rm_network_tool"}
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.100.0"
uvicorn = "0.23.1"
azure-identity = "1.13.0"
azure-mgmt-resource = "23.0.0"
azure-mgmt-compute = "30.0.0"
azure-mgmt-network = "^28.1.0"
azure-mgmt-authorization = "3.0.0"
pydantic = "^2.7.2"
httpx = "^0.27.0"  # Allow versions >=0.27.0 to satisfy mcp[cli] requirement
rich = "13.4.2"
black = "^25.1.0"
redis = "^5.0.1"
anyio = "^4.9.0"
gunicorn = "^23.0.0"
requests = "^2.31.0"
pyyaml = "^6.0.2"
mcp = {extras = ["cli"], version = "^1.7.1"}
networkx = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
pytest-asyncio = "0.21.1"

[tool.poetry.scripts]
start-proxy = "azure_rm_proxy.app.main:start"
generate-test-harnesses = "azure_rm_proxy.tools.generate_test_harnesses:main"
azure-rm-proxy-client = "azure_rm_proxy.client.cli:app"
dump-infra-data = "azure_rm_proxy.client.dump_infra_data:fetch_and_save_infra_data"
azure-rm-client = "azure_rm_client.cmd:main"
azrmc = "azure_rm_client.cmd:main"
mcp-server = "azure_rm_mcp_server.main:start"
check-vm-connectivity = "azure_rm_network_tool.vm_connectivity:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
